Change arrow functions to normal functions
I wanted to declare the functions after all variables are initialized as
the script was relying on the fact that the declaration of the variable
`paintColor` will be hoisted and initialized before the function
`paintCell()` is called, which it was. But I felt that function hoisting
made more sense than variable hoisting.

Hence why the change of arrow functions to normal functions, as normal
functions are hoisted but arrow functions aren't.